Fixed .rootmap generation for ROOT-5 on OS-X
Filtering rlibmap output to replace ".so" with ".dylib" on OS-X.

rlibmap seems to use ".so" instead of ".dylib" in it's output when
started with "rlibmap -l SOME_LIB.dylib" (why!?) - this doesn't happen
for "rlibmap -l SOME_LIB.SOME_OTHER_EXTENSION". As a result, automatic
library loading in ROOT-5 fails, as CINT looks for an ".so" instead of
a ".dylib" file.